requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23
alembic==1.13.0
requests==2.31.0
telethon==1.34.0
cryptography==41.0.7
pydantic==2.5.0
